requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.0
aiosignal==1.3.2
attrs==25.3.0
bidict==0.23.1
certifi==2025.4.26
charset-normalizer==3.4.2
chess==1.11.2
click==8.2.1
frozenlist==1.6.0
h11==0.16.0
idna==3.10
multidict==6.4.4
propcache==0.3.1
pygame==2.6.1
python-chess==1.999
python-engineio==4.12.1
python-socketio==5.13.0
requests==2.32.3
simple-websocket==1.1.0
urllib3==2.4.0
uvicorn==0.34.2
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.20.0
